lib/ffmpeg: add d3d11va decoder test class
To use this with current vaapi-fits semantics/options,
we can set LIBVA_DRIVER_NAME=d3d11 to ensure the proper
caps file is loaded.  Also, we can use "--device d3d11va"
to ensure render_device is set accordingly in middleware
pipeline commands.

Obviously, d3d11 is not a vaapi driver... and using
LIBVA_DRIVER_NAME to set this is awkward.  So we will
want to improve upon this at some point.

Signed-off-by: U. Artie Eoff <ullysses.a.eoff@intel.com>