ffmpeg/vaapi/encode: remove ipbmode check
This check is not always correct and causes more problems
than not.  In some cases, if bframes is not specified then
the driver/middleware is free to decide a default which
does not always agree with this check logic.  For example,
AVC VDENC on DG2 enables bframes if left unspecified from
ffmpeg-vaapi command-line.

Signed-off-by: U. Artie Eoff <ullysses.a.eoff@intel.com>