ffmpeg: use Codec enum
Different and inconsistent hard-coded strings that represent
the same codec can introduce bugs and make it harder to maintain.
Thus, use the new common Codec enum to improve consistency and
maintenance problems.

The bitdepth in codec strings are overkill and unnecessary, so
that distinction is removed.

For transcode, currently only 8-bit is supported and assumed.
Additional bitdepth support may be added in the future using
a different logic.

Signed-off-by: U. Artie Eoff <ullysses.a.eoff@intel.com>